Deferred grammar loading - initial implementation
The shadow grammars are now partially synchronized with the live state on each
pause callback. Grammars' blobs are synchronized and rules are unloaded/loaded.

  - List synchronization is not yet implemented.
  - Errors are stored in the correct format for a response, but responses aren't
    yet implemented.
  - Rule exclusivity is not yet implemented.
  - JIT Pauses are not yet triggered - synchronization will only happen when
    Dragon decides to pause.

There are still some stability issues - Dragon will sometimes crash/freeze while
listening and there is still a long hang when the microphone is first turned on
after an injection.