Send mimic updates asynchronously
Mimic processing was incomplete. This is a restructure, rather than a completion
of the previous implementation.

Mimic updates issue a pause. We need to process the unpause messages before the
mimic can complete and issue its callback. If we try and wait for the mimic
callback once we issue it, it will deadlock, because we're issuing on the Uv
thread.

Instead, we send the result of a successfully queued mimic asynchronously, once
it completes. The method used is just a queue - we push mimics onto it when we
queue them, and we pop them off when we get a generic callback. Since we aren't
directly synchronizing, this could be unreliable.